+++ src/App.tsx
@@ -1,20 +1,13 @@
 import React, { useState, useEffect, useRef, useMemo, useCallback } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { Plus, RotateCcw, CheckCircle2, MoreHorizontal, X, Clock } from 'lucide-react';
+import { Plus, RotateCcw } from 'lucide-react';
 import ProgressRing from './components/ProgressRing/ProgressRing';
 import SessionCard from './components/SessionCard/SessionCard';
 import { Session, TimerState } from './types';
-import Button from './components/ui/Button';
-import { useAuth } from './context/AuthContext';
-import { signOut } from 'firebase/auth';
-import { auth } from './firebase';
+import Navbar from './components/Navbar';
+
 
 // --- Main App Component ---
 const App: React.FC = () => {
-  const navigate = useNavigate();
-  const user = useAuth();
-  const isLoggedIn = !!user;
-
   // State
   const [sessions, setSessions] = useState<Session[]>(() => {
     const stored = localStorage.getItem('sessions');
@@ -75,7 +68,7 @@ const App: React.FC = () => {
   });
 
   // Settings / Menu State
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
+  
   const [resetTime, setResetTime] = useState(() => {
     let stored = localStorage.getItem('resetTime');
     if (stored !== undefined && stored !== null){
@@ -299,92 +292,19 @@ const App: React.FC = () => {
     }]);
   };
 
-  const handleLogout = async () => {
-    await signOut(auth);
-    navigate("/login");
-  }
-
   // Derived State for UI
   const activeSessionTitle = sessions.find(s => s.id === activeSessionId)?.title || "Ready to Focus";
 
   return (
     <div className="min-h-screen bg-slate-50 text-slate-800 font-sans selection:bg-emerald-100">
-      
-      {/* Top Navigation / Header */}
-      <header className="bg-white border-b border-slate-200 sticky top-0 z-30 px-6 py-4 flex items-center justify-between shadow-sm">
-        <div className="flex items-center gap-3">
-           <div className="w-10 h-10 bg-gradient-to-br from-emerald-400 to-teal-600 rounded-xl flex items-center justify-center text-white shadow-emerald-200 shadow-lg">
-             <CheckCircle2 size={24} />
-           </div>
-           <div>
-             <h1 className="font-bold text-xl tracking-tight text-slate-900">Focus Flow</h1>
-             <p className="text-xs text-slate-500 font-medium flex items-center gap-1">
-               {activeSessionId ? (
-                 <span className="flex items-center gap-1 text-emerald-600 animate-pulse">
-                   <span className="w-2 h-2 rounded-full bg-emerald-500"></span> Focusing on {activeSessionTitle}
-                 </span>
-               ) : "No active session"}
-             </p>
-           </div>
-        </div>
-
-        <div className="hidden md:flex items-center gap-6 relative">
-          {
-            !isLoggedIn ?
-            <>
-              <Button onClick={() => navigate("/login")} variant="outline">Login</Button>
-              <Button onClick={() => navigate("/register")} variant="secondary">Signup</Button>
-            </> :
-              <Button onClick={handleLogout} variant="ghost">Logout</Button>
-          }
-          
-          <div className="text-right">
-              <div className="text-xs text-slate-400 font-medium uppercase tracking-wider">Streak</div>
-              <div className="font-bold text-slate-700">{streak} Days</div>
-          </div>
-          <div className="h-8 w-px bg-slate-200"></div>
-          
-          <div className="relative">
-            <button
-                onClick={() => setIsMenuOpen(!isMenuOpen)}
-                className={`p-2 rounded-full transition-colors ${isMenuOpen ? 'bg-slate-100 text-slate-700' : 'hover:bg-slate-100 text-slate-500'}`}
-            >
-                <MoreHorizontal size={20} />
-            </button>
-            {isMenuOpen && (
-                <>
-                    <div className="fixed inset-0 z-40" onClick={() => setIsMenuOpen(false)}></div>
-                    <div className="absolute right-0 top-12 bg-white shadow-xl border border-slate-100 rounded-2xl p-5 w-72 z-50 animate-in fade-in zoom-in-95 duration-100">
-                          <div className="flex items-center justify-between mb-4">
-                            <h4 className="font-bold text-slate-800">Settings</h4>
-                            <button onClick={() => setIsMenuOpen(false)} className="text-slate-400 hover:text-slate-600">
-                                <X size={16} />
-                            </button>
-                          </div>
-                          
-                          <div className="space-y-4">
-                            <div className="flex flex-col gap-2">
-                                <label className="text-xs text-slate-500 font-bold uppercase tracking-wider flex items-center gap-1">
-                                  <Clock size={12} /> Auto-Reset Daily Goal
-                                </label>
-                                <p className="text-xs text-slate-400">Progress resets at this time daily.</p>
-                                <input 
-                                  type="time" 
-                                  value={resetTime} 
-                                  onChange={(e) => handleSetResetTime(e.target.value)}
-                                  className="border border-slate-200 bg-slate-50 rounded-lg p-2 text-sm focus:outline-none focus:border-emerald-500 focus:ring-2 focus:ring-emerald-100 w-full"
-                                />
-                            </div>
-                          </div>
-                    </div>
-                </>
-            )}
-          </div>
-        </div>
-      </header>
-
+      <Navbar
+        activeSessionTitle={activeSessionTitle}
+        activeSessionId={activeSessionId}
+        streak={streak}
+        resetTime={resetTime}
+        handleSetResetTime={handleSetResetTime}
+      />
       <main className="max-w-7xl mx-auto p-6 md:p-8">
-        
         {/* Hidden Audio Element */}
         <audio 
             ref={audioRef}
